vital/requirements/requirements.txt
pathos
pillow
numpy
scipy
seaborn
scikit-image
scikit-learn
pandas
h5py
PyYAML
tqdm
SimpleITK
torch>=1.7
torchvision>=0.8
pytorch-lightning>=1.2
torchmetrics
torchinfo
comet-ml>=3.3
Keras-Preprocessing
medpy
natsort
nibabel
albumentations
hydra-core
python-dotenv
scikit-video
xlrd
openpyxl
